requirements.txt
# Core
Django>=4.2.7,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.1
django-filter>=23.5
whitenoise>=6.6.0

# Database — install ONE of these:
mysqlclient>=2.2.0          # Option A: preferred (may need Visual C++ on Windows)
pymysql>=1.1.0              # Option B: pure-Python fallback (easier on Windows)

# Image handling
Pillow>=10.4.0

# AI
anthropic>=0.34.2
